Fix the bug for data normalization, the mean and std for mnist-m-train is [0.4549, 0.2208]

+++ src/utils/worker_utils.py
@@ -103,8 +103,7 @@ class MiniDataset(Dataset):
                  ]
             )
         elif self.data.ndim == 3:
-            # self.data = np.expand_dims(self.data, axis=1).astype(np.float64)
-            self.data = self.data.reshape((-1, 28, 28, 1)).astype("float32")
+            self.data = self.data.reshape((-1, 28, 28, 1)).astype("uint8")
             self.transform = transforms.Compose(
                 [transforms.ToTensor(),
                  transforms.Normalize((0.1307,), (0.3081,))
